Bugfixes, Destruction update
:General
- Updated conditions with Improved Chaos Bolt's cast time reduction

:Affliction
- Removed Doomguard/Infernal from Settings

:Destruction
- Removed setting for Minimum Chaos Bolt, was confusing people and you
could utterly break the CR with it.
- Improved the AOE rotation to be more responsive when Charred Remains
was selected as a talent.
- Improved the single-target rotation considerably, it should be a lot
more optimal now than before

:About Summon Pet
I'm having some slight issue getting Demonic Servitude to work correctly
as it doesn't seem to want to cast the regular or GoSup demons with the
ID's - will investigate further and come up with a fix asap.